SignObject tests
